HSEARCH-4806 Move Knn predicate impl to a correct package
